+++ serial/tools/hexlify_codec.py
@@ -18,6 +18,8 @@ Therefore decoding is binary to text and thus converting binary data to hex dump
 
 """
 
+from __future__ import absolute_import
+
 import codecs
 import serial